+++ code/packages/sdk/src/actions/ellipse-tool/ellipse-tool.ts
@@ -74,7 +74,9 @@ export class WeaveEllipseToolAction extends WeaveAction {
       }
     });
 
-    stage.on('pointerdown', () => {
+    stage.on('pointerdown', (e) => {
+      this.setTapStart(e);
+
       if (this.state === ELLIPSE_TOOL_STATE.ADDING) {
         this.creating = true;
 
@@ -82,7 +84,9 @@ export class WeaveEllipseToolAction extends WeaveAction {
       }
     });
 
-    stage.on('pointermove', () => {
+    stage.on('pointermove', (e) => {
+      if (!this.isPressed(e)) return;
+
       if (this.state === ELLIPSE_TOOL_STATE.DEFINING_SIZE) {
         this.moved = true;
 
@@ -90,7 +94,13 @@ export class WeaveEllipseToolAction extends WeaveAction {
       }
     });
 
-    stage.on('pointerup', () => {
+    stage.on('pointerup', (e) => {
+      const isTap = this.isTap(e);
+
+      if (isTap) {
+        this.moved = false;
+      }
+
       if (this.state === ELLIPSE_TOOL_STATE.DEFINING_SIZE) {
         this.creating = false;