+++ code/packages/sdk/src/actions/rectangle-tool/rectangle-tool.ts
@@ -73,7 +73,9 @@ export class WeaveRectangleToolAction extends WeaveAction {
       }
     });
 
-    stage.on('pointerdown', () => {
+    stage.on('pointerdown', (e) => {
+      this.setTapStart(e);
+
       if (this.state === RECTANGLE_TOOL_STATE.ADDING) {
         this.creating = true;
 
@@ -81,7 +83,9 @@ export class WeaveRectangleToolAction extends WeaveAction {
       }
     });
 
-    stage.on('pointermove', () => {
+    stage.on('pointermove', (e) => {
+      if (!this.isPressed(e)) return;
+
       if (this.state === RECTANGLE_TOOL_STATE.DEFINING_SIZE) {
         this.moved = true;
 
@@ -89,7 +93,13 @@ export class WeaveRectangleToolAction extends WeaveAction {
       }
     });
 
-    stage.on('pointerup', () => {
+    stage.on('pointerup', (e) => {
+      const isTap = this.isTap(e);
+
+      if (isTap) {
+        this.moved = false;
+      }
+
       if (this.state === RECTANGLE_TOOL_STATE.DEFINING_SIZE) {
         this.creating = false;