+++ code/packages/sdk/src/plugins/stage-zoom/types.ts
@@ -31,9 +31,19 @@ export type WeaveStageZoomPluginConfig = {
   zoomInertia: {
     friction: number;
     mouseWheelStep: number;
+    trackpadStep: number;
   };
 };
 
+type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends object
+    ? // eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
+      T[P] extends Function
+      ? T[P]
+      : DeepPartial<T[P]>
+    : T[P];
+};
+
 export type WeaveStageZoomPluginParams = {
-  config?: Partial<WeaveStageZoomPluginConfig>;
+  config?: DeepPartial<WeaveStageZoomPluginConfig>;
 };